+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Unreleased
 
+- Make `typing_extensions.TypeAliasType`'s `__module__` attribute writable.
+  Backport of CPython PR
+  [#149172](https://github.com/python/cpython/pull/149172).
+- Fix setting of `__required_keys__` and `__optional_keys__` when inheriting
+  keys with the same name.
+- Add support for `AsyncIterator`, `io.Reader`, `io.Writer` and `os.PathLike` protocols
+  as bases for other protocls.
 - Fix incorrect behaviour on Python 3.9 and Python 3.10 that meant that
   calling `isinstance` with `typing_extensions.Concatenate[...]` or
   `typing_extensions.Unpack[...]` as the first argument could have a different